refactor:ResultLoading.tsx를 EmojiAnimation.tsx와 TypingAnimation.tsx로 분리시킴
